+++ src/cmakeProject.ts
@@ -14,8 +14,8 @@ import { DirectoryContext } from '@cmt/workspace';
 import * as path from 'path';
 import * as vscode from 'vscode';
 import * as proc from '@cmt/proc';
-import * as api from './api';
-import { ExecutionOptions, ExecutionResult } from './api';
+import * as api from './legacyApi';
+import { ExecutionOptions, ExecutionResult } from './legacyApi';
 import { CodeModelContent } from '@cmt/drivers/codeModel';
 import { BadHomeDirectoryError } from '@cmt/drivers/cmakeServerClient';
 import { CMakeServerDriver, NoGeneratorError } from '@cmt/drivers/cmakeServerDriver';
