+++ src/drivers/cmakeServerDriver.ts
@@ -3,8 +3,8 @@ import { InputFileSet } from '@cmt/dirty';
 import { ConfigureTrigger } from '@cmt/cmakeProject';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import * as api from '@cmt/api';
-import { CacheEntryProperties, ExecutableTarget, RichTarget } from '@cmt/api';
+import * as api from '@cmt/legacyApi';
+import { CacheEntryProperties, ExecutableTarget, RichTarget } from '@cmt/legacyApi';
 import * as cache from '@cmt/cache';
 import * as cms from '@cmt/drivers/cmakeServerClient';
 import { CMakeDriver, CMakePreconditionProblemSolver } from '@cmt/drivers/cmakeDriver';
