+++ test/extension-tests/single-root-UI/test/envs-kits.test.ts
@@ -1,4 +1,4 @@
-import * as api from '@cmt/api';
+import * as api from '@cmt/legacyApi';
 import { CMakeCache } from '@cmt/cache';
 import { clearExistingKitConfigurationFile, DefaultEnvironment, expect, getFirstSystemKit } from '@test/util';
 import { fs } from '@cmt/pr';
