+++ test/extension-tests/single-root-UI/test/envs-presets.test.ts
@@ -1,4 +1,4 @@
-import * as api from '@cmt/api';
+import * as api from '@cmt/legacyApi';
 import { CMakeCache } from '@cmt/cache';
 import { DefaultEnvironment, expect } from '@test/util';
 import * as vscode from 'vscode';
