+++ test/extension-tests/successful-build/test/variable-substitution.test.ts
@@ -1,4 +1,4 @@
-import * as api from '@cmt/api';
+import * as api from '@cmt/legacyApi';
 import { CMakeCache } from '@cmt/cache';
 import { CMakeProject, ConfigureTrigger } from '@cmt/cmakeProject';
 import paths from '@cmt/paths';
