+++ test/extension-tests/successful-build/test/variant-envs.test.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-unused-expressions */
 
-import * as api from '@cmt/api';
+import * as api from '@cmt/legacyApi';
 import { CMakeCache } from '@cmt/cache';
 import { CMakeProject } from '@cmt/cmakeProject';
 import { clearExistingKitConfigurationFile, DefaultEnvironment, expect, getFirstSystemKit } from '@test/util';
